Implementing a range removal function
Given a vector and two indices, remove the elements in that vector starting at the first index up to the second index. In other words, the two indices divide the vector into three partitions. Remove the partition in the middle, and stitch the first and third partition together to form the resulting vector.